servers/ai/app/requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.2
asyncssh==2.21.1
